Conflicts:
	twitter4j-core/src/main/java/twitter4j/api/ListsResources.java

+++ twitter4j-async/src/main/java/twitter4j/AsyncTwitterImpl.java
@@ -1872,7 +1872,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void getUserListStatuses(final int listId, final Paging paging) {
+    public void getUserListStatuses(final long listId, final Paging paging) {
         getDispatcher().invokeLater(new AsyncTask(USER_LIST_STATUSES, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -1910,7 +1910,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void destroyUserListMember(final int listId, final long userId) {
+    public void destroyUserListMember(final long listId, final long userId) {
         getDispatcher().invokeLater(new AsyncTask(DESTROY_LIST_MEMBER, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -1926,7 +1926,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
     }
 
     @Override
-    public void deleteUserListMember(int listId, long userId) {
+    public void deleteUserListMember(long listId, long userId) {
         destroyUserListMember(listId, userId);
     }
 
@@ -2034,7 +2034,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void getUserListSubscribers(final int listId, final long cursor) {
+    public void getUserListSubscribers(final long listId, final long cursor) {
         getDispatcher().invokeLater(new AsyncTask(LIST_SUBSCRIBERS, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2072,7 +2072,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void createUserListSubscription(final int listId) {
+    public void createUserListSubscription(final long listId) {
         getDispatcher().invokeLater(new AsyncTask(SUBSCRIBE_LIST, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2110,7 +2110,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void showUserListSubscription(final int listId, final long userId) {
+    public void showUserListSubscription(final long listId, final long userId) {
         getDispatcher().invokeLater(new AsyncTask(CHECK_LIST_SUBSCRIPTION, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2148,7 +2148,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void destroyUserListSubscription(final int listId) {
+    public void destroyUserListSubscription(final long listId) {
         getDispatcher().invokeLater(new AsyncTask(UNSUBSCRIBE_LIST, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2187,7 +2187,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void createUserListMembers(final int listId, final long[] userIds) {
+    public void createUserListMembers(final long listId, final long[] userIds) {
         getDispatcher().invokeLater(new AsyncTask(CREATE_LIST_MEMBERS, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2203,7 +2203,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
     }
 
     @Override
-    public void addUserListMembers(int listId, long[] userIds) {
+    public void addUserListMembers(long listId, long[] userIds) {
         createUserListMembers(listId, userIds);
     }
 
@@ -2235,7 +2235,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void createUserListMembers(final int listId, final String[] screenNames) {
+    public void createUserListMembers(final long listId, final String[] screenNames) {
         getDispatcher().invokeLater(new AsyncTask(CREATE_LIST_MEMBERS, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2251,7 +2251,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
     }
 
     @Override
-    public void addUserListMembers(int listId, String[] screenNames) {
+    public void addUserListMembers(long listId, String[] screenNames) {
         createUserListMembers(listId, screenNames);
     }
 
@@ -2283,7 +2283,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void showUserListMembership(final int listId, final long userId) {
+    public void showUserListMembership(final long listId, final long userId) {
         getDispatcher().invokeLater(new AsyncTask(CHECK_LIST_MEMBERSHIP, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2321,7 +2321,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void getUserListMembers(final int listId, final long cursor) {
+    public void getUserListMembers(final long listId, final long cursor) {
         getDispatcher().invokeLater(new AsyncTask(LIST_MEMBERS, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2359,7 +2359,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void createUserListMember(final int listId, final long userId) {
+    public void createUserListMember(final long listId, final long userId) {
         getDispatcher().invokeLater(new AsyncTask(CREATE_LIST_MEMBER, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2375,7 +2375,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
     }
 
     @Override
-    public void addUserListMember(int listId, long userId) {
+    public void addUserListMember(long listId, long userId) {
         createUserListMember(listId, userId);
     }
 
@@ -2407,7 +2407,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void destroyUserList(final int listId) {
+    public void destroyUserList(final long listId) {
         getDispatcher().invokeLater(new AsyncTask(DESTROY_USER_LIST, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2445,7 +2445,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void updateUserList(final int listId, final String newListName, final boolean isPublicList, final String newDescription) {
+    public void updateUserList(final long listId, final String newListName, final boolean isPublicList, final String newDescription) {
         getDispatcher().invokeLater(new AsyncTask(UPDATE_USER_LIST, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
@@ -2502,7 +2502,7 @@ class AsyncTwitterImpl extends TwitterBaseImpl implements AsyncTwitter {
      * {@inheritDoc}
      */
     @Override
-    public void showUserList(final int listId) {
+    public void showUserList(final long listId) {
         getDispatcher().invokeLater(new AsyncTask(SHOW_USER_LIST, listeners) {
             @Override
             public void invoke(List<TwitterListener> listeners) throws TwitterException {
