+++ twitter4j-core/src/main/java/twitter4j/TwitterImpl.java
@@ -1086,7 +1086,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public ResponseList<Status> getUserListStatuses(int listId, Paging paging) throws TwitterException {
+    public ResponseList<Status> getUserListStatuses(long listId, Paging paging) throws TwitterException {
         return factory.createStatusList(get(conf.getRestBaseURL() + "lists/statuses.json"
                 , mergeParameters(paging.asPostParameterArray(Paging.SMCP, Paging.COUNT)
                 , new HttpParameter("list_id", listId))));
@@ -1119,14 +1119,14 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList destroyUserListMember(int listId, long userId) throws TwitterException {
+    public UserList destroyUserListMember(long listId, long userId) throws TwitterException {
         return factory.createAUserList(post(conf.getRestBaseURL() +
                 "lists/members/destroy.json",
                 new HttpParameter[]{new HttpParameter("list_id", listId), new HttpParameter("user_id", userId)}));
     }
 
     @Override
-    public UserList deleteUserListMember(int listId, long userId) throws TwitterException {
+    public UserList deleteUserListMember(long listId, long userId) throws TwitterException {
         return destroyUserListMember(listId, userId);
     }
 
@@ -1144,7 +1144,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList destroyUserListMember(int listId, String screenName) throws TwitterException {
+    public UserList destroyUserListMember(long listId, String screenName) throws TwitterException {
       return factory.createAUserList(post(conf.getRestBaseURL() +
           "lists/members/destroy.json", new HttpParameter[]{new HttpParameter("list_id", listId), 
         new HttpParameter("screen_name", screenName)}));
@@ -1170,7 +1170,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList destroyUserListMembers(int listId, String[] screenNames) throws TwitterException {
+    public UserList destroyUserListMembers(long listId, String[] screenNames) throws TwitterException {
       return factory.createAUserList(post(conf.getRestBaseURL() +
           "lists/members/destroy_all.json", new HttpParameter[]{new HttpParameter("list_id", listId), 
         new HttpParameter("screen_name",  z_T4JInternalStringUtil.join(screenNames))}));
@@ -1180,7 +1180,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList destroyUserListMembers(int listId, long[] userIds) throws TwitterException {
+    public UserList destroyUserListMembers(long listId, long[] userIds) throws TwitterException {
       return factory.createAUserList(post(conf.getRestBaseURL() +
           "lists/members/destroy_all.json", new HttpParameter[]{new HttpParameter("list_id", listId), 
         new HttpParameter("user_id", z_T4JInternalStringUtil.join(userIds))}));
@@ -1257,7 +1257,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public PagableResponseList<User> getUserListSubscribers(int listId, long cursor) throws TwitterException {
+    public PagableResponseList<User> getUserListSubscribers(long listId, long cursor) throws TwitterException {
         return factory.createPagableUserList(get(conf.getRestBaseURL() + "lists/subscribers.json?list_id=" + listId + "&cursor=" + cursor));
     }
 
@@ -1285,7 +1285,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList createUserListSubscription(int listId) throws TwitterException {
+    public UserList createUserListSubscription(long listId) throws TwitterException {
         return factory.createAUserList(post(conf.getRestBaseURL() +
                 "lists/subscribers/create.json", new HttpParameter[]{new HttpParameter("list_id", listId)}));
     }
@@ -1316,7 +1316,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public User showUserListSubscription(int listId, long userId) throws TwitterException {
+    public User showUserListSubscription(long listId, long userId) throws TwitterException {
         return factory.createUser(get(conf.getRestBaseURL() +
                 "lists/subscribers/show.json?list_id=" + listId + "&user_id=" + userId));
     }
@@ -1344,7 +1344,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList destroyUserListSubscription(int listId) throws TwitterException {
+    public UserList destroyUserListSubscription(long listId) throws TwitterException {
         return factory.createAUserList(post(conf.getRestBaseURL() +
                 "lists/subscribers/destroy.json", new HttpParameter[]{new HttpParameter("list_id", listId)}));
     }
@@ -1372,14 +1372,14 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList createUserListMembers(int listId, long[] userIds) throws TwitterException {
+    public UserList createUserListMembers(long listId, long[] userIds) throws TwitterException {
         return factory.createAUserList(post(conf.getRestBaseURL() + "lists/members/create_all.json",
                 new HttpParameter[]{new HttpParameter("list_id", listId), new HttpParameter("user_id"
                         , z_T4JInternalStringUtil.join(userIds))}));
     }
 
     @Override
-    public UserList addUserListMembers(int listId, long[] userIds) throws TwitterException {
+    public UserList addUserListMembers(long listId, long[] userIds) throws TwitterException {
         return createUserListMembers(listId, userIds);
     }
 
@@ -1413,7 +1413,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList createUserListMembers(int listId, String[] screenNames) throws TwitterException {
+    public UserList createUserListMembers(long listId, String[] screenNames) throws TwitterException {
         return factory.createAUserList(post(conf.getRestBaseURL() +
                 "lists/members/create_all.json",
                 new HttpParameter[]{
@@ -1422,7 +1422,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
     }
 
     @Override
-    public UserList addUserListMembers(int listId, String[] screenNames) throws TwitterException {
+    public UserList addUserListMembers(long listId, String[] screenNames) throws TwitterException {
         return createUserListMembers(listId, screenNames);
     }
 
@@ -1458,7 +1458,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public User showUserListMembership(int listId, long userId) throws TwitterException {
+    public User showUserListMembership(long listId, long userId) throws TwitterException {
         return factory.createUser(get(conf.getRestBaseURL() +
                 "lists/members/show.json?list_id=" + listId + "&user_id=" + userId));
     }
@@ -1486,7 +1486,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public PagableResponseList<User> getUserListMembers(int listId
+    public PagableResponseList<User> getUserListMembers(long listId
             , long cursor) throws TwitterException {
         return factory.createPagableUserList(get(conf.getRestBaseURL() +
                 "lists/members.json?list_id=" + listId + "&cursor=" + cursor));
@@ -1515,14 +1515,14 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList createUserListMember(int listId, long userId) throws TwitterException {
+    public UserList createUserListMember(long listId, long userId) throws TwitterException {
         return factory.createAUserList(post(conf.getRestBaseURL() +
                 "lists/members/create.json", new HttpParameter[]{new HttpParameter("user_id", userId)
                 , new HttpParameter("list_id", listId)}));
     }
 
     @Override
-    public UserList addUserListMember(int listId, long userId) throws TwitterException {
+    public UserList addUserListMember(long listId, long userId) throws TwitterException {
         return null;  //To change body of implemented methods use File | Settings | File Templates.
     }
 
@@ -1556,7 +1556,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList destroyUserList(int listId) throws TwitterException {
+    public UserList destroyUserList(long listId) throws TwitterException {
         return factory.createAUserList(post(conf.getRestBaseURL() + "lists/destroy.json",
                 new HttpParameter[]{new HttpParameter("list_id", listId)}));
     }
@@ -1586,7 +1586,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList updateUserList(int listId, String newListName, boolean isPublicList, String newDescription) throws TwitterException {
+    public UserList updateUserList(long listId, String newListName, boolean isPublicList, String newDescription) throws TwitterException {
         return updateUserList(newListName, isPublicList, newDescription, new HttpParameter("list_id", listId));
     }
 
@@ -1642,7 +1642,7 @@ class TwitterImpl extends TwitterBaseImpl implements Twitter {
      * {@inheritDoc}
      */
     @Override
-    public UserList showUserList(int listId) throws TwitterException {
+    public UserList showUserList(long listId) throws TwitterException {
         return factory.createAUserList(get(conf.getRestBaseURL() + "lists/show.json?list_id=" + listId));
     }
